+++ examples/cvode/fcmix_serial/fcvDiurnal_kry.f
@@ -166,7 +166,7 @@ C
 C
       INTEGER MX, MY
 C The following declaration specification should match C type long int.
-      INTEGER*4 IPAR(*)
+      INTEGER*8 IPAR(*)
       DOUBLE PRECISION RPAR(*)
 C
       INTEGER MM, JY, JX, P_IPP, P_BD, P_P
@@ -246,7 +246,7 @@ C
 C
       DOUBLE PRECISION T, U(2,*), UDOT(2,*), RPAR(*)
 C The following declaration specification should match C type long int.
-      INTEGER*4 IPAR(*)
+      INTEGER*8 IPAR(*)
       INTEGER IER
 C
       INTEGER ILEFT, IRIGHT
@@ -353,7 +353,7 @@ C
       INTEGER IER, JOK, JCUR
       DOUBLE PRECISION T, U(2,*), FU(*), GAMMA, H
 C The following declaration specification should match C type long int.
-      INTEGER*4 IPAR(*)
+      INTEGER*8 IPAR(*)
       DOUBLE PRECISION RPAR(*), V1(*), V2(*), V3(*)
 C
       INTEGER MX, MY, MM, P_IPP, P_BD, P_P
@@ -415,7 +415,7 @@ C
 C
       INTEGER IER, LR
 C The following declaration specification should match C type long int.
-      INTEGER*4 IPAR(*)
+      INTEGER*8 IPAR(*)
       DOUBLE PRECISION T, U(*), FU(*), R(*), Z(2,*)
       DOUBLE PRECISION GAMMA, DELTA, RPAR(*)
       DOUBLE PRECISION VTEMP(*)
